Make Repository inherit from Librarian#Repository and introduces the information related to the module directory under the git checkout to allow dependency resolution
