+++ lib/librarian/puppet/source/git.rb
@@ -1,55 +1,8 @@
 require 'librarian/source/git'
 require 'librarian/puppet/source/local'
+require 'librarian/puppet/source/repository'
 
 module Librarian
-  module Source
-    class Git
-      class Repository
-        def hash_from(remote, reference)
-          branch_names = remote_branch_names[remote]
-          if branch_names.include?(reference)
-            reference = "#{remote}/#{reference}"
-          end
-
-          command = %W(rev-parse #{reference}^{commit} --quiet)
-          run!(command, :chdir => true).strip
-        end
-
-        # Naming this method 'version' causes an exception to be raised.
-        def module_version
-          return '0.0.1' unless modulefile?
-
-          metadata  = ::Puppet::ModuleTool::Metadata.new
-          ::Puppet::ModuleTool::ModulefileReader.evaluate(metadata, modulefile)
-
-          metadata.version
-        end
-
-        def dependencies
-          return {} unless modulefile?
-
-          metadata = ::Puppet::ModuleTool::Metadata.new
-
-          ::Puppet::ModuleTool::ModulefileReader.evaluate(metadata, modulefile)
-
-          metadata.dependencies.inject({}) do |h, dependency|
-            name = dependency.instance_variable_get(:@full_module_name)
-            version = dependency.instance_variable_get(:@version_requirement)
-            h.update(name => version)
-          end
-        end
-
-        def modulefile
-          File.join(path, 'Modulefile')
-        end
-
-        def modulefile?
-          File.exists?(modulefile)
-        end
-      end
-    end
-  end
-
   module Puppet
     module Source
       class Git < Librarian::Source::Git
@@ -106,6 +59,12 @@ module Librarian
 
         def forge_source
           Forge.from_lock_options(environment, :remote=>"http://forge.puppetlabs.com")
+        end
+
+        def repository
+          @repository ||= begin
+            Repository.new(environment, repository_cache_path, @path)
+          end
         end
 
       end
